Continued work on refs #3 and ref #4 (dashboard)

+++ app/View/Elements/Project/list.ctp
@@ -15,13 +15,13 @@
 
 if (isset($projects)){
     foreach ($projects as $project): ?>
-        <div class="span4">
+        <? if (!isset($nospan)){?> <div class="span4"><?}?>
             <div class="well project-well">
-                <h3 class="project-title"><?=$this->Html->link($project['Project']['name'], array('action' => '.', 'project' => $project['Project']['name']), array('class' => 'project-link'))?>
+                <h3 class="project-title"><?=$this->Html->link($project['Project']['name'], array('controller' => 'projects', 'action' => '.', 'project' => $project['Project']['name']), array('class' => 'project-link'))?>
                 <span style="float: right;"><?= $this->Bootstrap->icon((($project['Project']['public']) ? 'globe' : 'lock'), 'black') ?></span></h3>
                 <p class="project-desc"><?= $this->Text->truncate($project['Project']['description'], 100)?></p>
                 <p class="project-time">Last Modified: <?=$this->Time->timeAgoInWords($project['Project']['modified'])?></p>
             </div>
-        </div>
+        <? if (!isset($nospan)){?> </div><?}?>
     <?php endforeach;
 } ?>
